Q&A_Chatbot/requirements.txt
langchain
langchain-openai
langchain_community
python-dotenv
langchain_core
streamlit